// src/types/tweetdeck.ts
export interface Column {
  clear(): void;
  model: ColumnModel;
  ui: ColumnUiState;
  _feeds: FeedState[];
  filters: any[];
  description: string;
  state: string;
  updateArray: ColumnChirp[];
  updateIndex: UpdateIndex;
  highWaterMarkChirp: ColumnChirp;
  highWaterMarkSeenChirp: ColumnChirp;
  isTrackingRealtime: boolean;
  isFetchingUpwardUpdates: boolean;
  isFetchingOlderUpdates: boolean;
  doUpwardScrollDeferreds: any[];
  numNewPushedChirps: number;
  wasPushedGap: boolean;
  scribedImpressionIDs: ScribedImpressionIDs;
  _isUpdating: boolean;
  feedSubscriptions: Subscriptions;
  deleteSubscriptions: Subscriptions;
  visible: boolean;
  visibility: Visibility;
  temporary: boolean;
  detailViewComponent: null | {
    chirp: TweetDeckChirp;
    mainChirp: TweetDeckChirp;
    repliesTo?: {
      repliesTo: TweetDeckChirp[];
    };
    replies?: {
      replies: TweetDeckChirp[];
    };
  };
  UPWARD_SCROLL_CHIRP_BLOCK_SIZE: number;
  INFINITE_SCROLL_CHIRP_BLOCK_SIZE: number;
  TARGET_COLUMN_CHIRP_LIMIT: number;
  RETRY_FILL_TIMEOUT: number;
  COLUMN_MINIMALIST_TWEET_TPM_THRESHOLD: number;
  CARDS_TPM_THRESHOLD: number;
  streamRate: number;
  streamRateEvent: string;
  tpm: number;
  tpmUpdateEvent: string;
  tpmCounter: unknown;
  tpmIncrementEvent: string;
  STREAM_ITEM_CONTENT_SELECTOR: string;
  STREAM_ITEM_SELECTOR: string;
  miscSubscriptions: Array<Array<null | string>>;
  chirpsWithPlayingGifs: any[];
  visibleChirpsEvent: string;
}

interface FeedState {
  state: FeedInternalState;
  privateState: FeedPrivateState;
  deltaQ: any[];
  stateCache: null;
  managed: boolean;
}

interface FeedPrivateState {
  key: string;
  apiid: string;
}

interface FeedInternalState {
  type: string;
  service: string;
  accountkey: string;
  metadata: unknown;
  mtime: Date;
}

interface Subscriptions {
  [k: string]: Array<null | string>;
}

interface ColumnChirp {
  account: ChirpAccount;
  targetTweet: TweetDeckChirp;
  sourceUser: User;
  created: Date;
  maxPosition: string;
  minPosition: string;
  id: string;
  text: string;
  htmlText: string;
  apiBounds: APIBounds;
  sortIndex: HighWaterMarkChirpSortIndex;
  _btd: Btd;
  chirpType: string;
  _hasAnimatedGif: boolean;
  messages?: TweetDeckChirp[];
  quotedTweet?: TweetDeckChirp;
  retweetedStatus?: TweetDeckChirp;
}

interface Btd {
  chirpKey: string;
  columnKey: string;
}

interface ChirpAccount {
  state: ChirpAccountState;
  privateState: ChirpAccountPrivateState;
  managed: boolean;
}

interface ChirpAccountPrivateState {
  isPrivate: boolean;
  verified: boolean;
  updated: number;
  key: string;
}

interface ChirpAccountState {
  type: string;
  oauth_token: string;
  username: string;
  name: string;
  profileImageURL: string;
  userId: string;
  require_some_consent: boolean;
}

interface APIBounds {
  [k: string]: ApiBound;
}

interface ApiBound {
  lower: string;
  upper: number;
}

interface HighWaterMarkChirpSortIndex {
  value: number;
  type: string;
}

interface User {
  account: ChirpAccount;
  id: string;
  screenName: string;
  profileURL: string;
  profileImageURL: string;
  avatarSuffixRegex: unknown;
  name: string;
  emojifiedName: string;
  location: string;
  description: string;
  entities: SourceUserEntities;
  joinedDate: Date;
  friendsCount: number;
  listedCount: number;
  followersCount: number;
  statusesCount: number;
  url: string;
  isProtected: boolean;
  isVerified: boolean;
  isTranslator: boolean;
  isBadged: boolean;
  following: boolean;
  lang: null;
  profileColor: string;
  bannerUrl: string;
  bannerUrlSmall: string;
  _profileBannerURL: string;
}

interface SourceUserEntities {
  url: UrlEntity;
  description: UrlEntity;
}

interface UrlEntity {
  urls: TwitterUrl[];
}

interface TwitterUrl {
  url: string;
  expanded_url: string;
  display_url: string;
  indices: number[];
}

export interface TweetDeckChirp {
  account: ChirpAccount;
  id: string;
  sortIndex: TargetTweetSortIndex;
  user: User;
  sourceNoHTML: string;
  sourceUrl: string;
  prettyRetweetCount: boolean;
  prettyLikeCount: boolean;
  prettyReplyCount: boolean;
  withPrettyEngagements: boolean;
  inReplyToUserID: string;
  inReplyToScreenName: string;
  isFavorite: boolean;
  isRetweeted: boolean;
  retweetCount: number;
  likeCount: number;
  replyCount: number;
  inReplyToID: string | number | null;
  selfThreadId: boolean;
  text: string;
  entities: TweetEntities;
  isQuoteStatus: boolean;
  quotedTweet?: TweetDeckChirp;
  targetTweet?: TweetDeckChirp;
  retweet?: TweetDeckChirp;
  retweetedStatus?: TweetDeckChirp;
  quotedTweetMissing: boolean;
  _media: any[];
  lang: string;
  place: unknown;
  created: Date;
  conversationMuted: boolean;
  htmlText: string;
  chirpType?: string;
  _btd?: {
    chirpKey: string;
    columnKey: string;
  };
  renderInMediaGallery(): string;
}

export interface TweetHashtag {
  indices: [number, number];
  text: string;
}

export interface TweetEntities {
  hashtags: Twitter.HashtagEntity[];
  urls: Twitter.UrlEntity[];
  user_mentions: Twitter.UserMentionEntity[];
  cashtags: any[];
  media: Twitter.MediaEntity[];
  convertedToUTF16: boolean;
}

interface TargetTweetSortIndex {
  value: string;
  type: string;
}

interface ColumnModel {
  state: ColumnModelState;
  privateState: FeedPrivateState;
  deltaQ: any[];
  stateCache: null;
  managed: boolean;
}

interface ColumnModelState {
  mtime: Date;
  type: string;
  settings: {
    media_preview_size: TweetDeckColumnMediaPreviewSizesEnum;
  };
  feedkeys: string[];
}

interface ScribedImpressionIDs {
  cache: ScribedImpressionIDsCache;
  head: Head;
  tail: Head;
  cacheSize: number;
  maxSize: number;
}

interface ScribedImpressionIDsCache {
  [k: string]: Head;
}

interface Head {
  key: string;
  value: boolean;
  next: null;
  previous: null;
}

interface ColumnUiState {
  chirpContainerSelector: string;
  infiniteSpinnerSelector: string;
  globalTpmSlidingLimit: number;
  columnTpmSlidingLimit: number;
  queuedSlidingLimit: number;
  chirpStaggeringInterval: number;
  $moreTweetsButtonContainer: MoreTweetsButtonContainer;
  fadeInClass: string;
  state: UIState;
  newTweetsTemplates: NewTweetsTemplates;
  moreTweetsState: MoreTweetsState;
  _$chirpContainer: ChirpContainer;
}

interface MoreTweetsButtonContainer {
  '0': unknown;
  length: number;
  prevObject: MoreTweetsButtonContainerPrevObject;
  context: Context;
  selector: string;
}
// ...
interface ChirpContainer {
  '0': unknown;
  length: number;
  prevObject: ChirpContainerPrevObject;
  context: Context;
  selector: string;
}

interface ChirpContainerPrevObject {
  '0': Context;
  context: Context;
  length: number;
}

interface MoreTweetsState {
  current: string;
  transition: null;
}

interface NewTweetsTemplates {
  default: string;
  'plural tweet': string;
  'singular tweet': string;
  'plural notification': string;
  'singular notification': string;
  'plural activity': string;
  'singular activity': string;
  'plural follower': string;
  'singular follower': string;
  'plural message': string;
  'singular message': string;
}

interface UIState {
  columnKey: string;
  $noResults: null;
  sliding: boolean;
  slidingQueue: any[];
}

interface UpdateIndex {
  [k: string]: ColumnChirp & TweetDeckChirp;
}

interface Visibility {
  columnWidth: number;
  visibleWidth: number;
  visibleHeight: number;
  visibleFraction: number;
}